Global cache configuration option
