+++ src/lightspeed_evaluation/core/constants.py
@@ -50,6 +50,10 @@ DISTANCE_MEASURE_MAP = {
     SIMILARITY_JARO_WINKLER: DistanceMeasure.JARO_WINKLER,
 }
 
+# Cache configuration
+DEFAULT_CACHE_BASE_DIR = ".caches"
+DEFAULT_API_CACHE_SUBDIR = "api"
+DEFAULT_LLM_CACHE_SUBDIR = "llm"
 
 # API Constants
 DEFAULT_API_BASE = "http://localhost:8080"
@@ -57,7 +61,6 @@ DEFAULT_API_VERSION = "v1"
 DEFAULT_API_TIMEOUT = 300
 DEFAULT_ENDPOINT_TYPE = "streaming"
 SUPPORTED_ENDPOINT_TYPES = ["streaming", "query", "infer"]
-DEFAULT_API_CACHE_DIR = ".caches/api_cache"
 
 DEFAULT_API_NUM_RETRIES = 3
 
@@ -75,11 +78,9 @@ DEFAULT_SSL_CERT_FILE = None
 DEFAULT_LLM_TEMPERATURE = 0.0
 DEFAULT_LLM_MAX_TOKENS = 512
 DEFAULT_LLM_RETRIES = 3
-DEFAULT_LLM_CACHE_DIR = ".caches/llm_cache"
 
 DEFAULT_EMBEDDING_PROVIDER = "openai"
 DEFAULT_EMBEDDING_MODEL = "text-embedding-3-small"
-DEFAULT_EMBEDDING_CACHE_DIR = ".caches/embedding_cache"
 
 DEFAULT_OUTPUT_DIR = "./eval_output"
 DEFAULT_BASE_FILENAME = "evaluation"
